Fixed debugger to easier debug one cpu at a time
Previously it would show a dump of whatever device was next scheduled
to step, if it was Debuggable, but now each debuggable device has a
flag for whether that device is being debugging, and a system flag
to disable it entirely.  When the system flag is set, it will try
to enable debugging on the device labelled "cpu".  I need to also add
a way of setting a breakpoint on a named device, which will enable
debugging of that device