Fixed bug in genesis controllers with incorrect counter
The TH counter in the genesis controller code was increasing each
time the TH bit changed state instead of only in one direction.
Surprisingly it worked before, and fixing it made it not work in
Sonic 2, but after adding the 1.5ms timer to reset the TH counter,
it worked again.  I guess Sonic2 only reads the 3 buttons instead
of all 6.  It should now work with both 3 and 6 button games